core: Move token events into their own module
