--- core/src/token/supply_management.rs
@@ -1,13 +1,0 @@
-// This Source Code Form is subject to the terms of the Mozilla Public
-// License, v. 2.0. If a copy of the MPL was not distributed with this
-// file, You can obtain one at http://mozilla.org/MPL/2.0/.
-//
-// Copyright (c) DUSK NETWORK. All rights reserved.
-
-/// Events emitted by supply management transactions.
-pub mod events {
-    /// The topic of the mint event.
-    pub const MINT_TOPIC: &str = "mint";
-    /// The topic of the burn event.
-    pub const BURN_TOPIC: &str = "burn";
-}
